+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Fix
 
 - Status and method are properties of hash;
+- `submit()` params clone;
 - `submit()` action type check.
 
 ## [1.0.2] - 2023-05-15